+++ homeassistant/components/telegram_client/entity.py
@@ -4,26 +4,26 @@ from __future__ import annotations
 
 from homeassistant.components.binary_sensor import BinarySensorEntityDescription
 from homeassistant.components.sensor import SensorEntityDescription
-from homeassistant.helpers.entity import Entity
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN
-from .device import TelegramClientDevice
+from .coordinator import TelegramClientCoordinator
 
 
-class TelegramClientEntity(Entity):
+class TelegramClientEntity(CoordinatorEntity):
     """Telegram client entity."""
 
     _attr_has_entity_name = True
 
     def __init__(
         self,
-        device: TelegramClientDevice,
+        coordinator: TelegramClientCoordinator,
         entity_description: BinarySensorEntityDescription | SensorEntityDescription,
     ) -> None:
         """Initialize the entity."""
-        self._device = device
+        super().__init__(coordinator)
         self.entity_description = entity_description
-        self._attr_device_info = device.device_info
+        self._attr_device_info = coordinator.device_info
         self._attr_unique_id = (
-            f"{DOMAIN}_{device.entry.unique_id}_{entity_description.key}"
+            f"{DOMAIN}_{coordinator.unique_id}_{entity_description.key}"
         )
